requirements.txt
pytorch_lightning
tensorboard
pygame-ce